added some code from http://ml-ants.googlecode.com/hg-history/8f438bb37142cf2c6592b04dd205d1f7b2fe39d3/src/bot/ExploreBot.py for diffusion around obstacles.  It's very fast and works a bit better than ours, but it's not totally right, and there are some bugs (negative values), but it works, and worth looking into more

+++ src/Birds/GameState.py
@@ -13,7 +13,7 @@ class GameState():
 	def __init__(self, mapFile):
 		"""Constructs a new GameState object"""
                 # load obstacle array from map file
-                Metric.obstacleAry = np.loadtxt(mapFile, dtype='c').astype('float')
+                Metric.obstacleAry = np.loadtxt(mapFile, dtype='c').astype('int')
                 # how many columns and rows are in this map?  (it is assumed that maps are rectangular)
                 # instantiate EntityGroups
                 self.entityGroups["Bird"] = EntityGroup(self, "Bird", [Metric("attract",  0.3, 30),
